pyproject.toml
[tool.poetry]
name = "squad-admin-configurator-discord"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "3.8.5"
py-cord = "2.5.0"
python-dateutil = "2.8.2"
jinja2 = "3.1.3"
toml = "0.10.2"

[tool.poetry.group.dev.dependencies]
aiohttp = "3.8.5"
py-cord = "2.5.0"
python-dateutil = "2.8.2"
jinja2 = "3.1.3"
toml = "0.10.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
